+++ seed/go-sdk/path-parameters/inline-path-parameters/user/client.go
@@ -31,45 +31,6 @@ func NewClient(opts ...option.RequestOption) *Client {
 	}
 }
 
-func (c *Client) GetOrganization(
-	ctx context.Context,
-	organizationId string,
-	opts ...option.RequestOption,
-) (*pathparametersgo.Organization, error) {
-	options := core.NewRequestOptions(opts...)
-	baseURL := internal.ResolveBaseURL(
-		options.BaseURL,
-		c.baseURL,
-		"",
-	)
-	endpointURL := internal.EncodeURL(
-		baseURL+"/user/organizations/%v",
-		organizationId,
-	)
-	headers := internal.MergeHeaders(
-		c.header.Clone(),
-		options.ToHeader(),
-	)
-
-	var response *pathparametersgo.Organization
-	if err := c.caller.Call(
-		ctx,
-		&internal.CallParams{
-			URL:             endpointURL,
-			Method:          http.MethodGet,
-			Headers:         headers,
-			MaxAttempts:     options.MaxAttempts,
-			BodyProperties:  options.BodyProperties,
-			QueryParameters: options.QueryParameters,
-			Client:          options.HTTPClient,
-			Response:        &response,
-		},
-	); err != nil {
-		return nil, err
-	}
-	return response, nil
-}
-
 func (c *Client) GetUser(
 	ctx context.Context,
 	request *pathparametersgo.GetUsersRequest,
@@ -82,47 +43,8 @@ func (c *Client) GetUser(
 		"",
 	)
 	endpointURL := internal.EncodeURL(
-		baseURL+"/user/users/%v",
-		request.UserId,
-	)
-	headers := internal.MergeHeaders(
-		c.header.Clone(),
-		options.ToHeader(),
-	)
-
-	var response *pathparametersgo.User
-	if err := c.caller.Call(
-		ctx,
-		&internal.CallParams{
-			URL:             endpointURL,
-			Method:          http.MethodGet,
-			Headers:         headers,
-			MaxAttempts:     options.MaxAttempts,
-			BodyProperties:  options.BodyProperties,
-			QueryParameters: options.QueryParameters,
-			Client:          options.HTTPClient,
-			Response:        &response,
-		},
-	); err != nil {
-		return nil, err
-	}
-	return response, nil
-}
-
-func (c *Client) GetOrganizationUser(
-	ctx context.Context,
-	request *pathparametersgo.GetOrganizationUserRequest,
-	opts ...option.RequestOption,
-) (*pathparametersgo.User, error) {
-	options := core.NewRequestOptions(opts...)
-	baseURL := internal.ResolveBaseURL(
-		options.BaseURL,
-		c.baseURL,
-		"",
-	)
-	endpointURL := internal.EncodeURL(
-		baseURL+"/user/organizations/%v/users/%v",
-		request.OrganizationId,
+		baseURL+"/%v/user/%v",
+		request.TenantId,
 		request.UserId,
 	)
 	headers := internal.MergeHeaders(
@@ -161,7 +83,8 @@ func (c *Client) SearchUsers(
 		"",
 	)
 	endpointURL := internal.EncodeURL(
-		baseURL+"/user/users/%v/search",
+		baseURL+"/%v/user/%v/search",
+		request.TenantId,
 		request.UserId,
 	)
 	queryParams, err := internal.QueryValues(request)
@@ -177,53 +100,6 @@ func (c *Client) SearchUsers(
 	)
 
 	var response []*pathparametersgo.User
-	if err := c.caller.Call(
-		ctx,
-		&internal.CallParams{
-			URL:             endpointURL,
-			Method:          http.MethodGet,
-			Headers:         headers,
-			MaxAttempts:     options.MaxAttempts,
-			BodyProperties:  options.BodyProperties,
-			QueryParameters: options.QueryParameters,
-			Client:          options.HTTPClient,
-			Response:        &response,
-		},
-	); err != nil {
-		return nil, err
-	}
-	return response, nil
-}
-
-func (c *Client) SearchOrganizations(
-	ctx context.Context,
-	organizationId string,
-	request *pathparametersgo.SearchOrganizationsRequest,
-	opts ...option.RequestOption,
-) ([]*pathparametersgo.Organization, error) {
-	options := core.NewRequestOptions(opts...)
-	baseURL := internal.ResolveBaseURL(
-		options.BaseURL,
-		c.baseURL,
-		"",
-	)
-	endpointURL := internal.EncodeURL(
-		baseURL+"/user/organizations/%v/search",
-		organizationId,
-	)
-	queryParams, err := internal.QueryValues(request)
-	if err != nil {
-		return nil, err
-	}
-	if len(queryParams) > 0 {
-		endpointURL += "?" + queryParams.Encode()
-	}
-	headers := internal.MergeHeaders(
-		c.header.Clone(),
-		options.ToHeader(),
-	)
-
-	var response []*pathparametersgo.Organization
 	if err := c.caller.Call(
 		ctx,
 		&internal.CallParams{