+++ seed/go-sdk/path-parameters/no-custom-config/dynamic-snippets/example4/snippet.go
@@ -8,10 +8,11 @@ import (
 
 func do() () {
     client := client.NewClient()
-    client.User.SearchOrganizations(
+    client.User.SearchUsers(
         context.TODO(),
-        "organizationId",
-        &fern.SearchOrganizationsRequest{
+        "tenantId",
+        "userId",
+        &fern.SearchUsersRequest{
             Limit: fern.Int(
                 1,
             ),