+++ seed/ts-sdk/path-parameters/src/serialization/resources/user/client/index.ts
@@ -1,2 +1,0 @@
-export * as searchUsers from "./searchUsers";
-export * as searchOrganizations from "./searchOrganizations";